test/inspector/heap-profiler/collect-garbage-expected.txt

Tests collectGarbage.

Running test: testConsoleRetainingPath
WeakRef state: WeakRef is cleared after GC.
